refactor: 유저 DTO 패키지 이동 및 전달 파라미터 변경
