+++ src/test/java/coffeemeet/server/common/fixture/dto/SignupDtoFixture.java
@@ -3,8 +3,8 @@ package coffeemeet.server.common.fixture.dto;
 import static org.instancio.Select.field;
 
 import coffeemeet.server.interest.domain.Keyword;
-import coffeemeet.server.user.dto.SignupHttpDto;
-import coffeemeet.server.user.dto.SignupHttpDto.Request;
+import coffeemeet.server.user.controller.dto.SignupHttpDto;
+import coffeemeet.server.user.controller.dto.SignupHttpDto.Request;
 import java.util.List;
 import org.instancio.Instancio;
